Replace hardcoded level datatype with appropriate type

+++ python/cuda_cccl/tests/parallel/examples/histogram/histogram_object.py
@@ -24,8 +24,8 @@ def histogram_object_example():
     # Configure histogram with 5 bins from 0 to 10
     num_levels = 6  # 5 bins = 6 levels
     h_num_output_levels = np.array([num_levels], dtype="int32")
-    h_lower_level = np.array([0.0], dtype="float64")
-    h_upper_level = np.array([10.0], dtype="float64")
+    h_lower_level = np.array([0.0], dtype="float32")
+    h_upper_level = np.array([10.0], dtype="float32")
 
     d_histogram = cp.zeros(num_levels - 1, dtype="int32")
